Added Item Preview and Bulk Import/Export
Items now show in the party list preview next to species and level.
To bulk import, select a folder containing numbered .trainer files. Does not require a .trainer file for every trainer.
To bulk export, select the folder to put all of the .trainer files. All trainers will be exported.